+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- introduce a new command `bit diff` to show the files diff for modified components
+
 ## [0.12.11] - 2018-04-10
 
 ### New